reorganize deprecated funcs
