+++ dynamo/preprocessing/preprocessor_utils.py
@@ -120,77 +120,6 @@ def clip_by_perc(layer_mat):
     return
 
 
-def calc_mean_var_dispersion_general_mat(
-    data_mat: Union[np.ndarray, csr_matrix], axis: int = 0
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """calculate mean, variance, and dispersion of a matrix.
-
-    Args:
-        data_mat: the matrix to be evaluated, either a ndarray or a scipy sparse matrix.
-        axis: the axis along which calculation is performed. Defaults to 0.
-
-    Returns:
-        A tuple (mean, var, dispersion) where mean is the mean of the array along the given axis, var is the variance of
-        the array along the given axis, and dispersion is the dispersion of the array along the given axis.
-    """
-
-    if not issparse(data_mat):
-        return calc_mean_var_dispersion_ndarray(data_mat, axis)
-    else:
-        return calc_mean_var_dispersion_sparse(data_mat, axis)
-
-
-def calc_mean_var_dispersion_ndarray(data_mat: np.ndarray, axis: int = 0) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """calculate mean, variance, and dispersion of a non-sparse matrix.
-
-    Args:
-        data_mat: the matrix to be evaluated.
-        axis: the axis along which calculation is performed. Defaults to 0.
-
-    Returns:
-        A tuple (mean, var, dispersion) where mean is the mean of the array along the given axis, var is the variance of
-        the array along the given axis, and dispersion is the dispersion of the array along the given axis.
-    """
-
-    # per gene mean, var and dispersion
-    mean = np.nanmean(data_mat, axis=axis).flatten()
-
-    # <class 'anndata._core.views.ArrayView'> has bug after using operator "==" (e.g. mean == 0), which changes mean.
-    mean = np.array(mean)
-    mean[mean == 0] += 1e-7  # prevent division by zero
-    var = np.nanvar(data_mat, axis=axis)
-    dispersion = var / mean
-    return mean.flatten(), var.flatten(), dispersion.flatten()
-
-
-def calc_mean_var_dispersion_sparse(sparse_mat: csr_matrix, axis: int = 0) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """calculate mean, variance, and dispersion of a matrix.
-
-    Args:
-        sparse_mat: the sparse matrix to be evaluated.
-        axis: the axis along which calculation is performed. Defaults to 0.
-
-    Returns:
-        A tuple (mean, var, dispersion), where mean is the mean of the array along the given axis, var is the variance
-        of the array along the given axis, and dispersion is the dispersion of the array along the given axis.
-    """
-
-    sparse_mat = sparse_mat.copy()
-    nan_mask = get_nan_or_inf_data_bool_mask(sparse_mat.data)
-    temp_val = (sparse_mat != 0).sum(axis)
-    sparse_mat.data[nan_mask] = 0
-    nan_count = temp_val - (sparse_mat != 0).sum(axis)
-
-    non_nan_count = sparse_mat.shape[axis] - nan_count
-    mean = (sparse_mat.sum(axis) / sparse_mat.shape[axis]).A1
-    mean[mean == 0] += 1e-7  # prevent division by zero
-
-    # same as numpy var behavior: denominator is N, var=(data_arr-mean)/N
-    var = np.power(sparse_mat - mean, 2).sum(axis) / sparse_mat.shape[axis]
-    dispersion = var / mean
-    return np.array(mean).flatten(), np.array(var).flatten(), np.array(dispersion).flatten()
-
-
 def seurat_get_mean_var(
     X: Union[csr_matrix, np.ndarray],
     ignore_zeros: bool = False,
