+++ dynamo/preprocessing/utils.py
@@ -178,103 +178,6 @@ def compute_gene_exp_fraction(X: scipy.sparse.spmatrix, threshold: float = 0.001
 
 
 # ---------------------------------------------------------------------------------------------------
-# implmentation of Cooks' distance (but this is for Poisson distribution fitting)
-
-# https://stackoverflow.com/questions/47686227/poisson-regression-in-statsmodels-and-r
-
-# from __future__ import division, print_function
-
-# https://stats.stackexchange.com/questions/356053/the-identity-link-function-does-not-respect-the-domain-of-the-gamma-
-# family
-def _weight_matrix(fitted_model: sm.Poisson) -> np.ndarray:
-    """Calculates weight matrix in Poisson regression.
-
-    Args:
-        fitted_model: a fitted Poisson model
-
-    Returns:
-        A diagonal weight matrix in Poisson regression.
-    """
-
-    return np.diag(fitted_model.fittedvalues)
-
-
-def _hessian(X: np.ndarray, W: np.ndarray) -> np.ndarray:
-    """Hessian matrix calculated as -X'*W*X.
-
-    Args:
-        X: the matrix of covariates.
-        W: the weight matrix.
-
-    Returns:
-        The result Hessian matrix.
-    """
-
-    return -np.dot(X.T, np.dot(W, X))
-
-
-def _hat_matrix(X: np.ndarray, W: np.ndarray) -> np.ndarray:
-    """Calculate hat matrix = W^(1/2) * X * (X'*W*X)^(-1) * X'*W^(1/2)
-
-    Args:
-        X: the matrix of covariates.
-        W: the diagonal weight matrix
-
-    Returns:
-        The result hat matrix
-    """
-
-    # W^(1/2)
-    Wsqrt = W ** (0.5)
-
-    # (X'*W*X)^(-1)
-    XtWX = -_hessian(X=X, W=W)
-    XtWX_inv = np.linalg.inv(XtWX)
-
-    # W^(1/2)*X
-    WsqrtX = np.dot(Wsqrt, X)
-
-    # X'*W^(1/2)
-    XtWsqrt = np.dot(X.T, Wsqrt)
-
-    return np.dot(WsqrtX, np.dot(XtWX_inv, XtWsqrt))
-
-
-def cook_dist(model: sm.Poisson, X: np.ndarray, good: npt.ArrayLike) -> np.ndarray:
-    """calculate Cook's distance
-
-    Args:
-        model: a fitted Poisson model.
-        X: the matrix of covariates.
-        good: the dispersion table for MSE calculation.
-
-    Returns:
-        The result Cook's distance.
-    """
-
-    # Weight matrix
-    W = _weight_matrix(model)
-
-    # Hat matrix
-    H = _hat_matrix(X, W)
-    hii = np.diag(H)  # Diagonal values of hat matrix # fit.get_influence().hat_matrix_diag
-
-    # Pearson residuals
-    r = model.resid_pearson
-
-    # Cook's distance (formula used by R = (res/(1 - hat))^2 * hat/(dispersion * p))
-    # Note: dispersion is 1 since we aren't modeling overdispersion
-
-    resid = good.disp - model.predict(good)
-    rss = np.sum(resid**2)
-    MSE = rss / (good.shape[0] - 2)
-    # use the formula from: https://www.mathworks.com/help/stats/cooks-distance.html
-    cooks_d = r**2 / (2 * MSE) * hii / (1 - hii) ** 2  # (r / (1 - hii)) ** 2 *  / (1 * 2)
-
-    return cooks_d
-
-
-# ---------------------------------------------------------------------------------------------------
 # preprocess utilities
 def unique_var_obs_adata(adata: anndata.AnnData) -> anndata.AnnData:
     """Function to make the obs and var attribute's index unique
